Working Bitfinex Websocket API (#6)

+++ Crux/Extend.cs
@@ -12,9 +12,9 @@ namespace Crux
             return Math.Sqrt(values.Average(v => Math.Pow(v - avg, 2)));
         }
 
-        public static int UnixTimestamp(this DateTime time)
+        public static long UnixTimestamp(this DateTime time)
         {
-            return (int)(DateTime.UtcNow.Subtract(new DateTime(1970, 1, 1))).TotalMilliseconds;
+            return (long)(DateTime.UtcNow.Subtract(new DateTime(1970, 1, 1))).TotalMilliseconds;
         }
     }
 }
